Project/backend/requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
idna==2.7
mysqlclient==1.3.13
oauthlib==2.1.0
python3-openid==3.1.0
pytz==2018.6
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24
